restructuration composant
